+++ frontend/src/home/components/background_3d/space_scence.jsx
