+++ frontend/src/home/pages/home_page.jsx
@@ -1,28 +1,27 @@
 // 메인 페이지 화면
-// tailwind css 사용
+// 라이브러리
 import { useEffect, useState } from 'react';
+import { Canvas } from '@react-three/fiber';
 
 // 컴포넌트
-import ProfileCard from "../../components/home/layout/profile";
-import SideWidget from "../../components/home/layout/side_widget";
-import DashBoard from "../../components/home/layout/dash_board";
-import StartButton from "../../components/home/layout/start_button";
+import ProfileCard from "../components/layout/profile";
+import SideWidget from "../components/layout/side_widget";
+import DashBoard from "../components/layout/dash_board";
+import StartButton from "../components/layout/start_button";
 import CharacterGroup from '../../components/characters/character_group';
+import Test3d from "../components/background_3d/space_scence"
 
 // 사운드
 import { playAudio, stopAudio } from '../../utils/sound';
 import { useSound } from '../../context/sound_context';
 import backgroundMusic from '../../assets/sounds/track1.mp3';
 
-import { Canvas } from '@react-three/fiber';
-import Test3d from "../../components/home/background_3d/space_scence"
-
+import { useUserCharacters }  from '../hooks/user_character';
 
 function HomePage() {
 
     const {toggledSound, setToggledSound} = useSound(); // 배경음
-    // 사용자가 가진 캐릭터 id
-    const [characterUserGetMap, setCharacterUserGetMap] = useState([]);
+    const characterUserGetMap = useUserCharacters(); // 캐릭터 훅
 
     useEffect(()=>{
         // 배경음이 true일 때만 소리재생
@@ -33,37 +32,6 @@ function HomePage() {
         }
 
     },[toggledSound]);
-
-      useEffect(() => {
-        const fetchData = async () => {
-          const token = sessionStorage.getItem("accessToken");
-
-          const [ownedRes, allRes] = await Promise.all([
-            fetch("http://localhost:8000/characters/my-characters", {
-              headers: { Authorization: `Bearer ${token}`, }
-            }),
-            fetch("http://localhost:8000/characters"),
-          ]);
-
-          if(!ownedRes.ok || !allRes.ok){
-            console.error("인증 실패 또는 서버 오류");
-            return;
-          }
-
-          const owned = await ownedRes.json();
-          const all = await allRes.json();
-
-          // 필요한 데이터만 뽑기
-          const ownedCharacters = all.filter(c =>
-            owned.some(o => o.character_id === c.character_id)
-          );
-
-          setCharacterUserGetMap(ownedCharacters);
-        };
-
-        fetchData();
-      }, []);
-
   
   return (
